Added non-inner class: HtsGetResponse.
